requirements.txt
# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0
playwright>=1.40.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# AI/ML
google-generativeai>=0.3.0  # Gemini API

# Utilities
python-dotenv>=1.0.0
urllib3>=2.1.0

# Date/Time
python-dateutil>=2.8.0
pytz>=2023.3

# Logging & Monitoring
loguru>=0.7.2

# Data Processing
hashlib  # Built-in, for content hashing
json  # Built-in
uuid  # Built-in

# HTTP Headers & User Agents
fake-useragent>=1.4.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Rate Limiting
ratelimit>=2.2.1
